removed EcoCacheResolver and added EcoCacheResolver feature to PackageToWrapperCacheResolver
