+++ packages/js/client/src/__tests__/core/uri-resolution/histories/can resolve cache - 1.json
@@ -1,25 +1,25 @@
 [
-  "wrap://file/$root-wrapper-dir/simple/build => EcoCacheResolver => wrapper (wrap://file/$root-wrapper-dir/simple/build)",
+  "wrap://file/$root-wrapper-dir/simple/build => PackageToWrapperCacheResolver => wrapper (wrap://file/$root-wrapper-dir/simple/build)",
   [
     "wrap://file/$root-wrapper-dir/simple/build => UriResolverAggregator => package (wrap://file/$root-wrapper-dir/simple/build)",
     [
       "wrap://file/$root-wrapper-dir/simple/build => ExtendableUriResolver => package (wrap://file/$root-wrapper-dir/simple/build)",
       [
-        "wrap://ens/ipfs-resolver.polywrap.eth => EcoCacheResolver => wrapper (wrap://ens/ipfs-resolver.polywrap.eth)",
+        "wrap://ens/ipfs-resolver.polywrap.eth => PackageToWrapperCacheResolver => wrapper (wrap://ens/ipfs-resolver.polywrap.eth)",
         [
           "wrap://ens/ipfs-resolver.polywrap.eth => UriResolverAggregator => package (wrap://ens/ipfs-resolver.polywrap.eth)",
           [
             "wrap://ens/ipfs-resolver.polywrap.eth => StaticResolver - Package (wrap://ens/ipfs-resolver.polywrap.eth) => package (wrap://ens/ipfs-resolver.polywrap.eth)"
           ]
         ],
-        "wrap://ens/ens-resolver.polywrap.eth => EcoCacheResolver => wrapper (wrap://ens/ens-resolver.polywrap.eth)",
+        "wrap://ens/ens-resolver.polywrap.eth => PackageToWrapperCacheResolver => wrapper (wrap://ens/ens-resolver.polywrap.eth)",
         [
           "wrap://ens/ens-resolver.polywrap.eth => UriResolverAggregator => package (wrap://ens/ens-resolver.polywrap.eth)",
           [
             "wrap://ens/ens-resolver.polywrap.eth => StaticResolver - Package (wrap://ens/ens-resolver.polywrap.eth) => package (wrap://ens/ens-resolver.polywrap.eth)"
           ]
         ],
-        "wrap://ens/fs-resolver.polywrap.eth => EcoCacheResolver => wrapper (wrap://ens/fs-resolver.polywrap.eth)",
+        "wrap://ens/fs-resolver.polywrap.eth => PackageToWrapperCacheResolver => wrapper (wrap://ens/fs-resolver.polywrap.eth)",
         [
           "wrap://ens/fs-resolver.polywrap.eth => UriResolverAggregator => package (wrap://ens/fs-resolver.polywrap.eth)",
           [
