+++ packages/js/client/src/__tests__/core/uri-resolution/histories/can resolve cache - 2.json
@@ -1,3 +1,3 @@
 [
-  "wrap://file/$root-wrapper-dir/simple/build => EcoCacheResolver (Cache) => wrapper (wrap://file/$root-wrapper-dir/simple/build)"
+  "wrap://file/$root-wrapper-dir/simple/build => PackageToWrapperCacheResolver (Cache) => wrapper (wrap://file/$root-wrapper-dir/simple/build)"
 ]
