+++ packages/js/client/src/__tests__/core/uri-resolution/histories/can resolve previously cached URI after redirecting by a URI resolver extension - 2.json
@@ -1,25 +1,25 @@
 [
-  "wrap://simple-redirect/$root-wrapper-dir/simple/build => EcoCacheResolver => uri (wrap://simple/$root-wrapper-dir/simple/build)",
+  "wrap://simple-redirect/$root-wrapper-dir/simple/build => PackageToWrapperCacheResolver => uri (wrap://simple/$root-wrapper-dir/simple/build)",
   [
     "wrap://simple-redirect/$root-wrapper-dir/simple/build => UriResolverAggregator => uri (wrap://simple/$root-wrapper-dir/simple/build)",
     [
       "wrap://simple-redirect/$root-wrapper-dir/simple/build => ExtendableUriResolver => uri (wrap://simple/$root-wrapper-dir/simple/build)",
       [
-        "wrap://ens/ipfs-resolver.polywrap.eth => EcoCacheResolver (Cache) => wrapper (wrap://ens/ipfs-resolver.polywrap.eth)",
-        "wrap://ens/ens-resolver.polywrap.eth => EcoCacheResolver (Cache) => wrapper (wrap://ens/ens-resolver.polywrap.eth)",
-        "wrap://ens/fs-resolver.polywrap.eth => EcoCacheResolver (Cache) => wrapper (wrap://ens/fs-resolver.polywrap.eth)",
-        "wrap://ens/http-resolver.polywrap.eth => EcoCacheResolver (Cache) => wrapper (wrap://ens/http-resolver.polywrap.eth)",
-        "wrap://ipfs/QmfRCVA1MSAjUbrXXjya4xA9QHkbWeiKRsT7Um1cvrR7FY => EcoCacheResolver (Cache) => wrapper (wrap://ipfs/QmfRCVA1MSAjUbrXXjya4xA9QHkbWeiKRsT7Um1cvrR7FY)",
-        "wrap://file/$root-wrapper-dir/simple-fs-resolver/build => EcoCacheResolver (Cache) => wrapper (wrap://file/$root-wrapper-dir/simple-fs-resolver/build)",
-        "wrap://file/$root-wrapper-dir/simple-redirect-resolver/build => EcoCacheResolver => wrapper (wrap://file/$root-wrapper-dir/simple-redirect-resolver/build)",
+        "wrap://ens/ipfs-resolver.polywrap.eth => PackageToWrapperCacheResolver (Cache) => wrapper (wrap://ens/ipfs-resolver.polywrap.eth)",
+        "wrap://ens/ens-resolver.polywrap.eth => PackageToWrapperCacheResolver (Cache) => wrapper (wrap://ens/ens-resolver.polywrap.eth)",
+        "wrap://ens/fs-resolver.polywrap.eth => PackageToWrapperCacheResolver (Cache) => wrapper (wrap://ens/fs-resolver.polywrap.eth)",
+        "wrap://ens/http-resolver.polywrap.eth => PackageToWrapperCacheResolver (Cache) => wrapper (wrap://ens/http-resolver.polywrap.eth)",
+        "wrap://ipfs/QmfRCVA1MSAjUbrXXjya4xA9QHkbWeiKRsT7Um1cvrR7FY => PackageToWrapperCacheResolver (Cache) => wrapper (wrap://ipfs/QmfRCVA1MSAjUbrXXjya4xA9QHkbWeiKRsT7Um1cvrR7FY)",
+        "wrap://file/$root-wrapper-dir/simple-fs-resolver/build => PackageToWrapperCacheResolver (Cache) => wrapper (wrap://file/$root-wrapper-dir/simple-fs-resolver/build)",
+        "wrap://file/$root-wrapper-dir/simple-redirect-resolver/build => PackageToWrapperCacheResolver => wrapper (wrap://file/$root-wrapper-dir/simple-redirect-resolver/build)",
         [
           "wrap://file/$root-wrapper-dir/simple-redirect-resolver/build => UriResolverAggregator => package (wrap://file/$root-wrapper-dir/simple-redirect-resolver/build)",
           [
             "wrap://file/$root-wrapper-dir/simple-redirect-resolver/build => ExtendableUriResolver => package (wrap://file/$root-wrapper-dir/simple-redirect-resolver/build)",
             [
-              "wrap://ens/ipfs-resolver.polywrap.eth => EcoCacheResolver (Cache) => wrapper (wrap://ens/ipfs-resolver.polywrap.eth)",
-              "wrap://ens/ens-resolver.polywrap.eth => EcoCacheResolver (Cache) => wrapper (wrap://ens/ens-resolver.polywrap.eth)",
-              "wrap://ens/fs-resolver.polywrap.eth => EcoCacheResolver (Cache) => wrapper (wrap://ens/fs-resolver.polywrap.eth)",
+              "wrap://ens/ipfs-resolver.polywrap.eth => PackageToWrapperCacheResolver (Cache) => wrapper (wrap://ens/ipfs-resolver.polywrap.eth)",
+              "wrap://ens/ens-resolver.polywrap.eth => PackageToWrapperCacheResolver (Cache) => wrapper (wrap://ens/ens-resolver.polywrap.eth)",
+              "wrap://ens/fs-resolver.polywrap.eth => PackageToWrapperCacheResolver (Cache) => wrapper (wrap://ens/fs-resolver.polywrap.eth)",
               "wrap://file/$root-wrapper-dir/simple-redirect-resolver/build => ResolverExtension (wrap://ens/fs-resolver.polywrap.eth) => package (wrap://file/$root-wrapper-dir/simple-redirect-resolver/build)"
             ]
           ]
@@ -28,5 +28,5 @@
       ]
     ]
   ],
-  "wrap://simple/$root-wrapper-dir/simple/build => EcoCacheResolver (Cache) => wrapper (wrap://simple/$root-wrapper-dir/simple/build)"
+  "wrap://simple/$root-wrapper-dir/simple/build => PackageToWrapperCacheResolver (Cache) => wrapper (wrap://simple/$root-wrapper-dir/simple/build)"
 ]
