+++ packages/js/client/src/__tests__/core/uri-resolution/histories/sanity.json
@@ -1,12 +1,12 @@
 [
-  "wrap://ens/uri.eth => EcoCacheResolver",
+  "wrap://ens/uri.eth => PackageToWrapperCacheResolver",
   [
     "wrap://ens/uri.eth => UriResolverAggregator",
     [
       "wrap://ens/uri.eth => StaticResolver - Miss",
       "wrap://ens/uri.eth => ExtendableUriResolver",
       [
-        "wrap://ens/ipfs-resolver.polywrap.eth => EcoCacheResolver => wrapper (wrap://ens/ipfs-resolver.polywrap.eth)",
+        "wrap://ens/ipfs-resolver.polywrap.eth => PackageToWrapperCacheResolver => wrapper (wrap://ens/ipfs-resolver.polywrap.eth)",
         [
           "wrap://ens/ipfs-resolver.polywrap.eth => UriResolverAggregator => package (wrap://ens/ipfs-resolver.polywrap.eth)",
           [
@@ -14,7 +14,7 @@
           ]
         ],
         "wrap://ens/uri.eth => ResolverExtension (wrap://ens/ipfs-resolver.polywrap.eth)",
-        "wrap://ens/ens-resolver.polywrap.eth => EcoCacheResolver => wrapper (wrap://ens/ens-resolver.polywrap.eth)",
+        "wrap://ens/ens-resolver.polywrap.eth => PackageToWrapperCacheResolver => wrapper (wrap://ens/ens-resolver.polywrap.eth)",
         [
           "wrap://ens/ens-resolver.polywrap.eth => UriResolverAggregator => package (wrap://ens/ens-resolver.polywrap.eth)",
           [
@@ -22,7 +22,7 @@
           ]
         ],
         "wrap://ens/uri.eth => ResolverExtension (wrap://ens/ens-resolver.polywrap.eth)",
-        "wrap://ens/fs-resolver.polywrap.eth => EcoCacheResolver => wrapper (wrap://ens/fs-resolver.polywrap.eth)",
+        "wrap://ens/fs-resolver.polywrap.eth => PackageToWrapperCacheResolver => wrapper (wrap://ens/fs-resolver.polywrap.eth)",
         [
           "wrap://ens/fs-resolver.polywrap.eth => UriResolverAggregator => package (wrap://ens/fs-resolver.polywrap.eth)",
           [
@@ -30,7 +30,7 @@
           ]
         ],
         "wrap://ens/uri.eth => ResolverExtension (wrap://ens/fs-resolver.polywrap.eth)",
-        "wrap://ens/http-resolver.polywrap.eth => EcoCacheResolver => wrapper (wrap://ens/http-resolver.polywrap.eth)",
+        "wrap://ens/http-resolver.polywrap.eth => PackageToWrapperCacheResolver => wrapper (wrap://ens/http-resolver.polywrap.eth)",
         [
           "wrap://ens/http-resolver.polywrap.eth => UriResolverAggregator => package (wrap://ens/http-resolver.polywrap.eth)",
           [
@@ -38,14 +38,14 @@
           ]
         ],
         "wrap://ens/uri.eth => ResolverExtension (wrap://ens/http-resolver.polywrap.eth)",
-        "wrap://ipfs/QmfRCVA1MSAjUbrXXjya4xA9QHkbWeiKRsT7Um1cvrR7FY => EcoCacheResolver => wrapper (wrap://ipfs/QmfRCVA1MSAjUbrXXjya4xA9QHkbWeiKRsT7Um1cvrR7FY)",
+        "wrap://ipfs/QmfRCVA1MSAjUbrXXjya4xA9QHkbWeiKRsT7Um1cvrR7FY => PackageToWrapperCacheResolver => wrapper (wrap://ipfs/QmfRCVA1MSAjUbrXXjya4xA9QHkbWeiKRsT7Um1cvrR7FY)",
         [
           "wrap://ipfs/QmfRCVA1MSAjUbrXXjya4xA9QHkbWeiKRsT7Um1cvrR7FY => UriResolverAggregator => package (wrap://ipfs/QmfRCVA1MSAjUbrXXjya4xA9QHkbWeiKRsT7Um1cvrR7FY)",
           [
             "wrap://ipfs/QmfRCVA1MSAjUbrXXjya4xA9QHkbWeiKRsT7Um1cvrR7FY => StaticResolver - Miss",
             "wrap://ipfs/QmfRCVA1MSAjUbrXXjya4xA9QHkbWeiKRsT7Um1cvrR7FY => ExtendableUriResolver => package (wrap://ipfs/QmfRCVA1MSAjUbrXXjya4xA9QHkbWeiKRsT7Um1cvrR7FY)",
             [
-              "wrap://ens/ipfs-resolver.polywrap.eth => EcoCacheResolver (Cache) => wrapper (wrap://ens/ipfs-resolver.polywrap.eth)",
+              "wrap://ens/ipfs-resolver.polywrap.eth => PackageToWrapperCacheResolver (Cache) => wrapper (wrap://ens/ipfs-resolver.polywrap.eth)",
               "wrap://ipfs/QmfRCVA1MSAjUbrXXjya4xA9QHkbWeiKRsT7Um1cvrR7FY => ResolverExtension (wrap://ens/ipfs-resolver.polywrap.eth) => package (wrap://ipfs/QmfRCVA1MSAjUbrXXjya4xA9QHkbWeiKRsT7Um1cvrR7FY)"
             ]
           ]
