+++ packages/js/uri-resolvers/src/__tests__/histories/resolve-ipfs-with-cache.json
@@ -1,3 +1,3 @@
 [
-  "wrap://ipfs/QmdEMfomFW1XqoxcsCEnhujn9ebQezUXw8pmwLtecyR6F6 => EcoCacheResolver (Cache) => wrapper (wrap://ipfs/QmdEMfomFW1XqoxcsCEnhujn9ebQezUXw8pmwLtecyR6F6)"
+  "wrap://ipfs/QmdEMfomFW1XqoxcsCEnhujn9ebQezUXw8pmwLtecyR6F6 => PackageToWrapperCacheResolver (Cache) => wrapper (wrap://ipfs/QmdEMfomFW1XqoxcsCEnhujn9ebQezUXw8pmwLtecyR6F6)"
 ]
