+++ packages/js/uri-resolvers/src/cache/PackageToWrapperCacheResolver.ts
@@ -1,5 +1,5 @@
 import { IWrapperCache } from "./IWrapperCache";
-import { UriResolver, UriResolutionResult, UriResolverLike } from "../helpers";
+import { UriResolutionResult, UriResolver, UriResolverLike } from "../helpers";
 
 import {
   IUriResolver,
@@ -10,27 +10,27 @@ import {
 } from "@polywrap/core-js";
 import { DeserializeManifestOptions } from "@polywrap/wrap-manifest-types-js";
 import { Result } from "@polywrap/result";
+import { Mutex, MutexInterface } from "async-mutex";
 
 export class PackageToWrapperCacheResolver<TError>
   implements IUriResolver<TError | Error> {
-  name: string;
+  private locks: Map<
+    string,
+    { mutex: Mutex; release?: MutexInterface.Releaser }
+  > = new Map();
 
   constructor(
     private _resolverToCache: IUriResolver<TError>,
     private _cache: IWrapperCache,
     private _options?: {
       deserializeManifestOptions?: DeserializeManifestOptions;
-      endOnRedirect?: boolean;
     }
   ) {}
 
   static from<TResolverError = unknown>(
     resolver: UriResolverLike,
     cache: IWrapperCache,
-    options?: {
-      deserializeManifestOptions?: DeserializeManifestOptions;
-      endOnRedirect?: boolean;
-    }
+    options?: { deserializeManifestOptions?: DeserializeManifestOptions }
   ): PackageToWrapperCacheResolver<TResolverError> {
     return new PackageToWrapperCacheResolver(
       UriResolver.from<TResolverError>(resolver),
@@ -39,13 +39,24 @@ export class PackageToWrapperCacheResolver<TError>
     );
   }
 
+  async isCached(uri: Uri): Promise<boolean> {
+    return Boolean(await this._cache.get(uri));
+  }
+
+  isLocked(uri: Uri): boolean {
+    return this.locks.has(uri.uri);
+  }
+
   async tryResolveUri(
     uri: Uri,
     client: CoreClient,
     resolutionContext: IUriResolutionContext
   ): Promise<Result<UriPackageOrWrapper, TError | Error>> {
+    await this.acquireMutex(uri);
+
     const wrapper = await this._cache.get(uri);
 
+    // return from cache if available
     if (wrapper) {
       const result = UriResolutionResult.ok(uri, wrapper);
 
@@ -57,42 +68,17 @@ export class PackageToWrapperCacheResolver<TError>
       return result;
     }
 
+    // resolve uri if not in cache
     const subContext = resolutionContext.createSubHistoryContext();
 
-    let result = await this._resolverToCache.tryResolveUri(
-      uri,
-      client,
-      subContext
-    );
+    let result: Result<
+      UriPackageOrWrapper,
+      TError | Error
+    > = await this._resolverToCache.tryResolveUri(uri, client, subContext);
 
     if (result.ok) {
-      if (result.value.type === "package") {
-        const wrapPackage = result.value.package;
-        const resolutionPath: Uri[] = subContext.getResolutionPath();
-
-        const createResult = await wrapPackage.createWrapper({
-          noValidate: this._options?.deserializeManifestOptions?.noValidate,
-        });
-
-        if (!createResult.ok) {
-          return createResult;
-        }
-
-        const wrapper = createResult.value;
-
-        for (const uri of resolutionPath) {
-          await this._cache.set(uri, wrapper);
-        }
-
-        result = UriResolutionResult.ok(result.value.uri, wrapper);
-      } else if (result.value.type === "wrapper") {
-        const wrapper = result.value.wrapper;
-        const resolutionPath: Uri[] = subContext.getResolutionPath();
-
-        for (const uri of resolutionPath) {
-          await this._cache.set(uri, wrapper);
-        }
-      }
+      result = await this.cacheResult(result, subContext);
+      this.cancelMutex(uri);
     }
 
     resolutionContext.trackStep({
@@ -102,5 +88,62 @@ export class PackageToWrapperCacheResolver<TError>
       description: "PackageToWrapperCacheResolver",
     });
     return result;
+  }
+
+  private async cacheResult(
+    result: Result<UriPackageOrWrapper, TError | Error>,
+    subContext: IUriResolutionContext
+  ): Promise<Result<UriPackageOrWrapper, TError | Error>> {
+    if (!result.ok || result.value.type === "uri") {
+      return result;
+    } else if (result.value.type === "package") {
+      const wrapPackage = result.value.package;
+      const resolutionPath: Uri[] = subContext.getResolutionPath();
+
+      const createResult = await wrapPackage.createWrapper({
+        noValidate: this._options?.deserializeManifestOptions?.noValidate,
+      });
+
+      if (!createResult.ok) {
+        return createResult;
+      }
+
+      const wrapper = createResult.value;
+
+      for (const uri of resolutionPath) {
+        await this._cache.set(uri, wrapper);
+      }
+
+      return UriResolutionResult.ok(result.value.uri, wrapper);
+    } else {
+      const wrapper = result.value.wrapper;
+      const resolutionPath: Uri[] = subContext.getResolutionPath();
+
+      for (const uri of resolutionPath) {
+        await this._cache.set(uri, wrapper);
+      }
+
+      return result;
+    }
+  }
+
+  private async acquireMutex(uri: Uri): Promise<void> {
+    const isCached = Boolean(await this._cache.get(uri));
+    if (!isCached) {
+      let lock = this.locks.get(uri.uri);
+      if (!lock) {
+        lock = { mutex: new Mutex() };
+        this.locks.set(uri.uri, lock);
+      }
+      lock.release = await lock.mutex.acquire().catch((_e) => undefined);
+    }
+  }
+
+  private cancelMutex(uri: Uri): void {
+    const uriStr = uri.uri;
+    const lock = this.locks.get(uriStr);
+    if (!lock) return;
+    lock.mutex.cancel();
+    this.locks.delete(uriStr);
   }
 }
