+++ packages/js/uri-resolvers/src/cache/index.ts
@@ -1,4 +1,3 @@
-export * from "./EcoCacheResolver";
 export * from "./PackageToWrapperCacheResolver";
 export * from "./IWrapperCache";
 export * from "./WrapperCache";
